+++ lib/src/item.dart
@@ -19,7 +19,7 @@ import 'package:flutter/cupertino.dart';
 /// Tab item used for [ConvexAppBar].
 class TabItem<T> {
   /// Tab text.
-  final String title;
+  final String? title;
 
   /// IconData or Image.
   ///
@@ -27,7 +27,7 @@ class TabItem<T> {
   final T icon;
 
   /// Optional if not provided ,[icon] is used.
-  final T activeIcon;
+  final T? activeIcon;
 
   /// Whether icon should blend with color.
   /// If [icon] is instance of [IconData] then blend is default to true, otherwise false
@@ -35,10 +35,10 @@ class TabItem<T> {
 
   /// Create item
   const TabItem({
-    this.title,
-    @required this.icon,
+    this.title = '',
+    required this.icon,
     this.activeIcon,
-    bool isIconBlend,
+    bool? isIconBlend,
   })  : assert(icon is IconData || icon is Widget,
             'TabItem only support IconData and Widget'),
         blend = isIconBlend ?? (icon is IconData);